Fix Node 11 builds (#81)
This changes our AppVeyor build config to use `Install-Product` instead of `Update-NodeJsInstallation`. It should speed up our builds by quite a bit since it's just switching the version to the latest one available on the image. This also resolves the issue we've been having with `npm` not being available as well.